Change f32 to f64 everywhere

+++ native/core/src/entropy.rs
@@ -9,12 +9,12 @@ use crate::{
     structs::{WordN, EntropiesData, Dictionary},
 };
 
-pub fn entropy(arr: Array1<f32>) -> f32 {
+pub fn entropy(arr: Array1<f64>) -> f64 {
     let arr = arr
         .into_iter()
         .filter(|&x| x > 0.)
-        .map(|x| x as f32)
-        .collect::<Array1<f32>>();
+        .map(|x| x as f64)
+        .collect::<Array1<f64>>();
 
     let arr_log = {
         let mut arr_log = arr.clone();
@@ -30,7 +30,7 @@ pub fn calculate_entropies<const N: usize>(
     dictionary: &Dictionary<N>,
     possible_answers: &Vec<WordN<char, N>>,
 ) -> Vec<EntropiesData<N>> {
-    let n = possible_answers.len() as f32;
+    let n = possible_answers.len() as f64;
 
     let guess_words = &dictionary.words;
     let trans = &dictionary.translator;
@@ -46,15 +46,15 @@ pub fn calculate_entropies<const N: usize>(
     let entropies = guess_words_iter
         .map(|guess| {
             let guess_b = trans.to_bytes(guess);
-            let mut guess_hints = FxHashMap::<_, f32>::default();
+            let mut guess_hints = FxHashMap::<_, f64>::default();
             for correct in possible_answers.iter() {
                 let mut left = ArrayVec::<_, N>::new();
                 let hints = algo::get_hints_with_work_array(&guess_b, correct, &mut left);
                 *guess_hints.entry(hints).or_default() += 1. / n;
             }
 
-            let probs = Array1::<f32>::from_vec(
-                guess_hints.values().map(|x| *x as f32).collect::<Vec<_>>(),
+            let probs = Array1::<f64>::from_vec(
+                guess_hints.values().map(|x| *x as f64).collect::<Vec<_>>(),
             );
             let entropy = entropy(probs);
